requirements.txt
flasgger==0.9.5
Flask==2.2.2
pyserial==3.5
requests==2.28.1
